+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 - `CollectionsPlume`:
    * add methods `duplicates()` and `listFilter()`
    * add an overload for `mapCapacity()`
+   * deprecated `noDuplicates()`; use `hasNoDuplicates()` instead
+   * deprecated `listFilter()`; use `filter()` instead
+- `ArraysPlume`:
+   * deprecated `noDuplicates()`; use `hasNoDuplicates()` instead
 
 ## 1.8.1 (2023-06-02)